+++ src/Utils/Logging.hpp
@@ -4,33 +4,7 @@
 #include <mutex>
 #include <source_location>
 #include <string>
-#include <sstream>
-
-class str
-{
-public:
-  template<typename ...VArgs>
-  static std::string agg(VArgs ...vargs)
-  {
-    std::ostringstream ss;
-    _agg(ss, vargs...);
-    return ss.str();
-  }
-
-private:
-  template<typename T, typename ...VArgs>
-  static void _agg(std::ostream& ss, T t, VArgs... vargs)
-  {
-    ss << t;
-    _agg(ss, vargs...);
-  }
-  
-  template<typename T>
-  static void _agg(std::ostream& ss, T t)
-  {
-    ss << t;
-  }
-};
+#include "Str.hpp"
 
 class Log
 {
